Undelete redundante codes
