+++ scripts/generate_code.py
@@ -14,71 +14,18 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import argparse
 import filecmp
 import glob
-import platform
 import shutil
 import subprocess
+import generate_grpc_examples
 from pathlib import Path
-
-parser = argparse.ArgumentParser()
-parser.add_argument(
-    "--flatc",
-    help="path of the Flat C compiler relative to the root directory",
-)
-parser.add_argument("--cpp-0x", action="store_true", help="use --cpp-std c++ox")
-parser.add_argument(
-    "--skip-monster-extra",
-    action="store_true",
-    help="skip generating tests involving monster_extra.fbs",
-)
-parser.add_argument(
-    "--skip-gen-reflection",
-    action="store_true",
-    help="skip generating the reflection.fbs files",
-)
-args = parser.parse_args()
-
-# Get the path where this script is located so we can invoke the script from
-# any directory and have the paths work correctly.
-script_path = Path(__file__).parent.resolve()
-
-# Get the root path as an absolute path, so all derived paths are absolute.
-root_path = script_path.parent.absolute()
-
-# Get the location of the flatc executable, reading from the first command line
-# argument or defaulting to default names.
-flatc_exe = Path(
-    ("flatc" if not platform.system() == "Windows" else "flatc.exe")
-    if not args.flatc
-    else args.flatc
-)
-
-# Find and assert flatc compiler is present.
-if root_path in flatc_exe.parents:
-    flatc_exe = flatc_exe.relative_to(root_path)
-flatc_path = Path(root_path, flatc_exe)
-assert flatc_path.exists(), "Cannot find the flatc compiler " + str(flatc_path)
+from util import flatc, root_path, tests_path, args, flatc_path
 
 # Specify the other paths that will be referenced
-tests_path = Path(root_path, "tests")
 swift_code_gen = Path(root_path, "tests/swift/tests/CodeGenerationTests")
 samples_path = Path(root_path, "samples")
 reflection_path = Path(root_path, "reflection")
-
-# Execute the flatc compiler with the specified parameters
-def flatc(options, schema, prefix=None, include=None, data=None, cwd=tests_path):
-    cmd = [str(flatc_path)] + options
-    if prefix:
-        cmd += ["-o"] + [prefix]
-    if include:
-        cmd += ["-I"] + [include]
-    cmd += [schema] if isinstance(schema, str) else schema
-    if data:
-        cmd += [data] if isinstance(data, str) else data
-    result = subprocess.run(cmd, cwd=str(cwd), check=True)
-
 
 # Generate the code for flatbuffers reflection schema
 def flatc_reflection(options, location, target):
@@ -560,50 +507,5 @@ flatc_annotate(
     schema="monster_test.fbs", include="include_test", data="monsterdata_test.mon"
 )
 
-# Generate grpc examples
-grpc_examples_path = Path(root_path, "grpc/examples")
-
-greeter_schema = Path(grpc_examples_path, "greeter.fbs")
-
-COMMON_ARGS = [
-    "--grpc",
-    "--bfbs-filenames",
-    str(grpc_examples_path),
-]
-
-flatc(
-    COMMON_ARGS
-    + [
-        "--go",
-    ],
-    schema=str(greeter_schema),
-    cwd=Path(grpc_examples_path, "go/greeter"),
-)
-
-flatc(
-    COMMON_ARGS
-    + [
-        "--python",
-    ],
-    schema=str(greeter_schema),
-    cwd=Path(grpc_examples_path, "python/greeter"),
-)
-
-flatc(
-    COMMON_ARGS
-    + [
-        "--swift",
-        "--gen-json-emit",
-    ],
-    schema=str(greeter_schema),
-    cwd=Path(grpc_examples_path, "swift/Greeter/Sources/Model"),
-)
-
-flatc(
-    COMMON_ARGS
-    + [
-        "--ts",
-    ],
-    schema=str(greeter_schema),
-    cwd=Path(grpc_examples_path, "ts/greeter/src"),
-)
+# Run the generate_grpc_examples script
+generate_grpc_examples.GenerateGRPCExamples()
